saving memory
Former-commit-id: bd3a312712deca8a7c0b7a9acf906f248c16d8b7

+++ COOP/iCache.h
@@ -3,7 +3,7 @@
 
 #include "object.h"
 
-typedef int MEM_SIZE_T;
+typedef unsigned short MEM_SIZE_T;
 
 DEF_CLASS(iCache);
 END_DEF(iCache);
